+++ HISTORY.md
@@ -1,6 +1,10 @@
 History
 =======
 
+## UNRELEASED
+
+- Add `fullPath` processed module field internally and to output reports.
+
 ## 4.1.1
 
 - BUG: A loader in the `identifier` field would incorrectly have all modules inferred "as a `node_modules` file", even if not. Implements a naive loader stripping heuristic to correctly assess if `node_modules` or real application source.